p4i integration Integrate 184589 from 4.0 to main: Fixes:    Floating point values for html tag attributes (like <table border=1.0) RevBy:    Andy AutoTest: #184588 Details:  As we support it in the rendering we might as well parse it correctly :)
[git-p4: depot-paths = "//depot/qt/main/": change = 184590]

+++ src/gui/text/qtexthtmlparser_p.h
@@ -177,14 +177,14 @@ struct QTextHtmlParserNode {
     QString anchorHref;
     QString anchorName;
     QString imageName;
-    int imageWidth;
-    int imageHeight;
+    qreal imageWidth;
+    qreal imageHeight;
     QTextLength width;
-    int tableBorder;
+    qreal tableBorder;
     int tableCellRowSpan;
     int tableCellColSpan;
-    int tableCellSpacing;
-    int tableCellPadding;
+    qreal tableCellSpacing;
+    qreal tableCellPadding;
 
     int cssBlockIndent;
     int cssListIndent;
